feat(fsm): add fsm application - separate FSM stack
