+++ mysite/ct/tests.py
@@ -7,10 +7,12 @@ Replace this with more appropriate tests for your application.
 
 from django.contrib.auth.models import User
 from django.test import TestCase
-from django.http import HttpResponseRedirect
 from ct.models import *
-from ct import views, fsm, ct_util
-import time, urllib
+from fsm.models import *
+from ct import views, ct_util
+import time
+import urllib
+
 
 class OurTestCase(TestCase):
     def check_post_get(self, url, postdata, urlTail, expected):
@@ -25,6 +27,7 @@ class OurTestCase(TestCase):
         self.assertContains(response, expected)
         return url
 
+
 class ConceptMethodTests(OurTestCase):
     def setUp(self):
         self.user = User.objects.create_user(username='jacob', email='jacob@_',
@@ -34,6 +37,7 @@ class ConceptMethodTests(OurTestCase):
                                              password='top_secret')
         self.unit = Unit(title='My Courselet', addedBy=self.user)
         self.unit.save()
+
     def test_sourceDB(self):
         'check wikipedia concept retrieval'
         c, lesson = Concept.get_from_sourceDB('New York City', self.user)
@@ -51,12 +55,14 @@ class ConceptMethodTests(OurTestCase):
         self.assertEqual(c2.pk, c.pk)
         self.assertEqual(l2.pk, lesson.pk)
         self.assertIn(c, list(Concept.search_text('new york')))
+
     def test_sourceDB_temp(self):
         'check wikipedia temporary document retrieval'
         lesson = Lesson.get_from_sourceDB('New York City', self.user,
                                           doSave=False)
         self.assertIn('City of New York', lesson.text) # got the text?
         self.assertEqual(Lesson.objects.count(), 0) # nothing saved?
+
     def test_wikipedia_view(self):
         'check wikipedia view and concept addition method'
         url = '/ct/teach/courses/1/units/%d/concepts/wikipedia/%s/' \
@@ -70,6 +76,7 @@ class ConceptMethodTests(OurTestCase):
         self.assertTrue(ul in UnitLesson.search_sourceDB('New York City')[0])
         self.assertTrue(ul in UnitLesson.search_sourceDB('New York City',
                                                          unit=self.unit)[0])
+
     def test_new_concept(self):
         'check standard creation of a concept bound to a UnitLesson'
         title = 'Important Concept'
@@ -92,6 +99,7 @@ class ConceptMethodTests(OurTestCase):
         self.assertEqual(lesson.kind, Lesson.ERROR_MODEL)
         ul = UnitLesson.objects.get(lesson=lesson)
         self.assertEqual(ul.kind, UnitLesson.MISUNDERSTANDS)
+
     def test_error_models(self):
         'check creation and copying of error models'
         concept = Concept.new_concept('big', 'idea', self.unit, self.user)
@@ -139,7 +147,7 @@ class ConceptMethodTests(OurTestCase):
         em, resols = ul.get_em_resolutions()
         resols = list(resols)
         self.assertEqual(resols, [resoUL, resoUL2])
-        
+
     def test_get_conceptlinks(self):
         'test ConceptLink creation and retrieval'
         concept = Concept.new_concept('bad', 'idea', self.unit, self.user)
@@ -166,7 +174,6 @@ class ConceptMethodTests(OurTestCase):
         self.assertEqual(distinct_subset([ul1, ul2, ul3]), [ul1, ul2])
 
 
-
 class LessonMethodTests(TestCase):
     def setUp(self):
         self.user = User.objects.create_user(username='jacob', email='jacob@_',
@@ -177,6 +184,7 @@ class LessonMethodTests(TestCase):
         self.ul.lesson.save()
         self.unit2 = Unit(title='My Courselet', addedBy=self.user)
         self.unit2.save()
+
     def test_creation_treeID(self):
         'treeID properly initialized to default?'
         lesson = Lesson(title='foo', text='bar', addedBy=self.user)
@@ -190,12 +198,13 @@ class LessonMethodTests(TestCase):
         self.assertTrue(len(results) >= 10)
         self.assertIn('New York City', [t[0] for t in results])
         self.assertEqual(len(results[0]), 3)
+
     def test_checkout(self):
         'check Lesson commit and checkout'
         self.assertFalse(self.ul.lesson.is_committed())
         self.ul.lesson.conceptlink_set.create(concept=self.ul.lesson.concept,
                                               addedBy=self.user)
-        ul2 = self.ul.copy(self.unit2, self.user) # copy to new unit
+        ul2 = self.ul.copy(self.unit2, self.user)  # copy to new unit
         self.assertTrue(self.ul.lesson.is_committed())
         self.assertEqual(self.ul.lesson, ul2.lesson)
         self.assertEqual(self.ul.lesson.changeLog, 'snapshot for fork by jacob')
@@ -216,33 +225,6 @@ class LessonMethodTests(TestCase):
         self.assertTrue(ul2b.lesson.is_committed())
 
 
-fsmDict = dict(name='test', title='try this')
-nodeDict = dict(START=dict(title='start here', path='ct:home',
-                           funcName='testme.START'),
-                MID=dict(title='in the middle', path='ct:about', doLogging=True),
-                END=dict(title='end here', path='ct:home'),
-    )
-edgeDict = (
-    dict(name='next', fromNode='START', toNode='END', title='go go go'),
-    dict(name='select_Lesson', fromNode='MID', toNode='MID', title='go go go'),
-    )
-
-def load_fsm2(username):
-    'load an FSM that displays a standard node page'
-    fsmDict = dict(name='test2', title='try this')
-    nodeDict = dict(START=dict(title='Welcome, Stranger', path='ct:fsm_node',
-                               description='Thanks for your valuable input!'),
-                    END=dict(title='end here', path='ct:home'),
-        )
-    edgeDict = (
-        dict(name='next', fromNode='START', toNode='END',
-             title='Write an Amazing Question',
-             description='''Your mission, should you choose to accept it,
-             is to write an amazing question that yields wonderful insights.'''),
-        )
-    return FSM.save_graph(fsmDict, nodeDict, edgeDict, username)
-
-    
 class FakeRequest(object):
     'trivial holder for request data to pass to test calls'
     def __init__(self, user, sessionDict=None, method='POST', dataDict=None,
@@ -257,6 +239,7 @@ class FakeRequest(object):
             dataDict = {}
         setattr(self, method, dataDict)
 
+
 def create_question_unit(user, utitle='Ask Me some questions',
                          qtitle='What is your quest?',
                          text="(That's a rather personal question.)"):
@@ -268,261 +251,28 @@ def create_question_unit(user, utitle='Ask Me some questions',
     ul = UnitLesson.create_from_lesson(question, unit, addAnswer=True,
                                        order='APPEND')
     return ul
-                    
-class FSMTests(OurTestCase):
-    def setUp(self):
-        self.user = User.objects.create_user(username='jacob', email='jacob@_',
-                                             password='top_secret')
-        # have to login or Django self.client.session storage won't work
-        self.client.login(username='jacob', password='top_secret')
-        self.course = Course(title='Great Course', description='the bestest',
-                             addedBy=self.user)
-        self.course.save()
-        self.unit = Unit(title='My Courselet', addedBy=self.user)
-        self.unit.save()
-        self.lesson = Lesson(title='Big Deal', text='very interesting info',
-                             addedBy=self.user)
-        self.lesson.save_root()
-        self.unitLesson = UnitLesson.create_from_lesson(self.lesson, self.unit,
-                                                        order='APPEND')
-        self.ulQ = create_question_unit(self.user)
-        self.ulQ2 = create_question_unit(self.user, 'Pretest', 'Scary Question',
-                                         'Tell me something.')
-    def test_load(self):
-        'check loading an FSM graph, and replacing it'
-        f = FSM.save_graph(fsmDict, nodeDict, edgeDict, 'jacob')
-        self.assertEqual(f.fsmnode_set.count(), 3)
-        self.assertEqual(f.startNode.name, 'START')
-        self.assertEqual(f.startNode.outgoing.count(), 1)
-        e = f.startNode.outgoing.all()[0]
-        self.assertEqual(e.name, 'next')
-        self.assertEqual(e.toNode.name, 'END')
-        f2 = FSM.save_graph(fsmDict, nodeDict, edgeDict, 'jacob') # replace
-        self.assertEqual(FSM.objects.get(pk=f.pk).name, 'testOLD') # renamed
-        self.assertNotEqual(f.startNode, f2.startNode)
-        self.assertEqual(f.startNode.name, f2.startNode.name)
-    def test_json_blob(self):
-        'check roundtrip dump/load via json blob data'
-        name, pk = dump_json_id(self.unit)
-        label, obj = load_json_id(name, pk)
-        self.assertEqual(self.unit, obj)
-        self.assertEqual(label, 'Unit')
-    def test_json_blob2(self):
-        'check roundtrip dump/load via named json blob data'
-        name, pk = dump_json_id(self.unit, 'fruity')
-        self.assertEqual(name, 'fruity_Unit_id')
-        label, obj = load_json_id(name, pk)
-        self.assertEqual(self.unit, obj)
-        self.assertEqual(label, 'fruity')
-    def test_json_blob3(self):
-        'check roundtrip dump/load via json blob string'
-        s = dump_json_id_dict(dict(fruity=self.unit))
-        d = load_json_id_dict(s)
-        self.assertEqual(d.items(), [('fruity', self.unit)])
-    def test_json_blob4(self):
-        'check roundtrip dump/load via db storage'
-        f = FSM.save_graph(fsmDict, nodeDict, edgeDict, 'jacob')
-        d = f.startNode.load_json_data()
-        self.assertEqual(d, {})
-        d['fruity'] = self.unit
-        d['anumber'] = 3
-        d['astring'] = 'jeff'
-        f.startNode.save_json_data(d)
-        node = FSMNode.objects.get(pk=f.startNode.pk)
-        d2 = node.load_json_data()
-        self.assertEqual(d2, {'fruity': self.unit, 'anumber': 3,
-                              'astring': 'jeff'})
-    def test_start(self):
-        'check basic startup of new FSM instance'
-        f = FSM.save_graph(fsmDict, nodeDict, edgeDict, 'jacob')
-        self.do_start(f)
-    def test_start2(self):
-        'check basic startup of new FSM instance using FSMSpecification'
-        from ct.fsm_plugin.testme import get_specs
-        spec = get_specs()[0]
-        f = spec.save_graph('jacob')
-        self.assertTrue(f.startNode.doLogging)
-        self.assertFalse(f.get_node('MID').doLogging)
-        fsmStack = self.do_start(f)
-        # test get_help() plugin functionality
-        request = FakeRequest(self.user, path='/ct/about/')
-        msg = fsmStack.state.fsmNode.get_help(fsmStack.state, request)
-        self.assertEqual(msg, 'here here!')
-        request = FakeRequest(self.user, path='/ct/courses/1/')
-        msg = fsmStack.state.fsmNode.get_help(fsmStack.state, request)
-        self.assertEqual(msg, 'there there')
-        request = FakeRequest(self.user)
-        msg = fsmStack.state.fsmNode.get_help(fsmStack.state, request)
-        self.assertEqual(msg, None)
-    def test_start3(self):
-        'check that FSMState saves unitLesson, select_ data, and logging'
-        f = FSM.save_graph(fsmDict, nodeDict, edgeDict, 'jacob')
-        fsmStack = self.do_start(f, unitLesson=self.unitLesson)
-        self.assertEqual(fsmStack.state.unitLesson, self.unitLesson)
-        request = FakeRequest(self.user, method='GET')
-        # check logging on the MID node
-        fsmStack.event(request, None) # send a render event to log
-        self.assertEqual(fsmStack.state.activity.fsmName, 'test')
-        ae = fsmStack.state.activityEvent
-        self.assertEqual(ae.nodeName, 'MID')
-        self.assertEqual(ae.unitLesson, self.unitLesson)
-        self.assertIsNotNone(ae.startTime)
-        self.assertIsNone(ae.endTime)
-        request = FakeRequest(self.user)
-        # now try a select_ event
-        fsmStack.event(request, 'select_Lesson', lesson=self.lesson)
-        self.assertEqual(fsmStack.state.get_data_attr('lesson'), self.lesson)
-        # check that exit from MID node was logged
-        self.assertIsNotNone(ae.endTime)
-        self.assertTrue(ae.endTime > ae.startTime)
-        self.assertEqual(ae.exitEvent, 'select_Lesson')
-        self.assertIsNone(fsmStack.state.activityEvent)
-    def do_start(self, f, **kwargs):
-        'run tests of basic startup of new FSM instance'
-        fsmData = dict(unit=self.unit, foo='bar')
-        request = FakeRequest(self.user)
-        fsmStack = fsm.FSMStack(request)
-        self.assertIsNone(fsmStack.state)
-        try:
-            result = fsmStack.push(request, 'invalid', stateData=fsmData, **kwargs)
-        except FSM.DoesNotExist:
-            pass
-        else:
-            raise AssertionError('failed to catch bad FSM query')
-        result = fsmStack.push(request, 'test', stateData=fsmData, **kwargs)
-        self.assertEqual(request.session['fsmID'], fsmStack.state.pk)
-        self.assertEqual(fsmStack.state.load_json_data(), fsmData)
-        self.assertEqual(fsmStack.state.fsmNode.name, 'MID')
-        self.assertEqual(fsmStack.state.fsmNode.path, 'ct:about')
-        self.assertEqual(fsmStack.state.fsmNode.get_path(fsmStack.state,
-                                                    request), '/ct/about/')
-        self.assertEqual(result, '/ct/about/')
-        return fsmStack
-    def test_trivial_plugin(self):
-        'check trivial plugin import and call'
-        f = FSM.save_graph(fsmDict, nodeDict, edgeDict, 'jacob')
-        request = FakeRequest(self.user)
-        fsmStack = fsm.FSMStack(request)
-        fsmStack.state = FSMState(user=self.user, fsmNode=f.startNode)
-        self.assertEqual(f.startNode.event(fsmStack, request, 'start'),
-                         '/ct/about/')
-        self.assertEqual(f.startNode.get_path(fsmStack.state, request),
-                         '/ct/some/where/else/')
-    def test_bad_funcName(self):
-        'check that FSM.save_graph() catches bad plugin funcName'
-        nodeDictBad = dict(START=dict(title='start here', path='ct:home',
-                           funcName='testme.invalid'))
-
-        try:
-            f = FSM.save_graph(fsmDict, nodeDictBad, (), 'jacob')
-        except AttributeError:
-            pass
-        else:
-            raise AssertionError('FSM.save_graph() failed to catch bad plugin funcName')
-    def test_bad_fsmID(self):
-        'make sure FSMStack silently handles bad fsmID'
-        request = FakeRequest(self.user, dict(fsmID=99))
-        fsmStack = fsm.FSMStack(request)
-        self.assertEqual(request.session, {})
-        self.assertIsNone(fsmStack.state)
-    def test_randomtrial(self):
-        'basic randomized trial'
-        self.assertEqual(self.ulQ.order, 0)
-        from ct.fsm_plugin.lessonseq import get_specs
-        f = get_specs()[0].save_graph(self.user.username) # load FSM spec
-        from ct.fsm_plugin.randomtrial import get_specs
-        f = get_specs()[0].save_graph(self.user.username) # load FSM spec
-        self.assertEqual(ActivityLog.objects.count(), 0)
-        fsmData = dict(testFSM='lessonseq', treatmentFSM='lessonseq',
-                       treatment1=self.ulQ.unit, treatment2=self.ulQ.unit,
-                       testUnit=self.ulQ2.unit, course=self.course)
-        request, fsmStack, result = self.get_fsm_request('randomtrial',
-                                        fsmData, dict(trialName='test'))
-        self.assertEqual(self.client.session['fsmID'], fsmStack.state.pk)
-        self.assertEqual(result, '/ct/nodes/%d/' % f.startNode.pk)
-        self.assertEqual(ActivityLog.objects.count(), 1)
-        # rt FSM start page
-        response = self.client.get(result)
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, 'answer a few preliminary questions')
-        url = '/ct/courses/%d/units/%d/lessons/%d/ask/' \
-              % (self.course.pk, self.ulQ2.unit.pk, self.ulQ2.pk)
-        self.check_post_get(result, dict(fsmtask='next'), url, 'Scary Question')
-        # pretest Q 
-        postdata = dict(text='i dunno', confidence=Response.GUESS)
-        url = self.check_post_get(url, postdata, '/assess/', 'write an answer')
-        # pretest assess
-        assessPOST = dict(selfeval=Response.CORRECT, status=DONE_STATUS,
-                        liked='')
-        url2 = '/ct/courses/%d/units/%d/lessons/%d/ask/' \
-              % (self.course.pk, self.ulQ.unit.pk, self.ulQ.pk)
-        self.check_post_get(url, assessPOST, url2, 'your quest')
-        # treatment Q
-        postdata = dict(text='i like rats', confidence=Response.GUESS)
-        url = self.check_post_get(url2, postdata, '/assess/', 'write an answer')
-        # treatment assess 
-        url2 = '/ct/courses/%d/units/%d/lessons/%d/ask/' \
-              % (self.course.pk, self.ulQ2.unit.pk, self.ulQ2.pk)
-        self.check_post_get(url, assessPOST, url2, 'Scary Question')
-        # posttest Q 
-        postdata = dict(text='i still dunno', confidence=Response.GUESS)
-        url = self.check_post_get(url2, postdata, '/assess/', 'write an answer')
-        # posttest assess
-        url2 = '/ct/courses/%d/units/%d/tasks/' \
-              % (self.course.pk, self.ulQ.unit.pk)
-        self.check_post_get(url, assessPOST, url2, 'Next Step to work on')
-        self.assertEqual(Response.objects.filter(activity__isnull=False,
-                                                 confidence=Response.GUESS).
-                         count(), 3) # check responses logged to RT activity
-
-    def test_slideshow(self):
-        'basic slide show FSM'
-        from ct.fsm_plugin.slideshow import get_specs
-        f = get_specs()[0].save_graph(self.user.username) # load FSM spec
-        fsmData = dict(unit=self.ulQ2.unit, course=self.course)
-        request, fsmStack, result = self.get_fsm_request('slideshow', fsmData)
-        # start page = question
-        response = self.client.get(result)
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, 'Scary Question')
-        # answer page
-        answer = self.ulQ2.get_answers()[0]
-        url = '/ct/courses/%d/units/%d/lessons/%d/' \
-              % (self.course.pk, self.ulQ2.unit.pk, answer.pk)
-        self.check_post_get(result, dict(fsmtask='next'), url, 'an answer')
-        # end of slide show should dump us on concepts page
-        url2 = '/ct/courses/%d/units/%d/concepts/' \
-              % (self.course.pk, self.ulQ2.unit.pk)
-        self.check_post_get(url, dict(fsmtask='next'), url2, 'Pretest')
-        
-    def get_fsm_request(self, fsmName, stateData, startArgs={}, **kwargs):
-        'create request, fsmStack and start specified FSM'
-        request = FakeRequest(self.user)
-        request.session = self.client.session
-        fsmStack = fsm.FSMStack(request)
-        result = fsmStack.push(request, fsmName, stateData, startArgs, **kwargs)
-        request.session.save()
-        return request, fsmStack, result
 
 
-        
 class ReversePathTests(TestCase):
     def test_home(self):
         'test trimming of args not needed for target'
         url = ct_util.reverse_path_args('ct:home',
                         '/ct/teach/courses/21/units/33/errors/2/')
         self.assertEqual(url, reverse('ct:home'))
+
     def test_ul_teach(self):
         'check proper extraction of args from path'
         url = ct_util.reverse_path_args('ct:ul_teach',
                         '/ct/teach/courses/21/units/33/errors/2/')
         self.assertEqual(url, reverse('ct:ul_teach', args=(21, 33, 2)))
+
     def test_ul_id(self):
         'test handling of ID kwargs'
         url = ct_util.reverse_path_args('ct:ul_teach',
                         '/ct/teach/courses/21/units/33/', ul_id=2)
         self.assertEqual(url, reverse('ct:ul_teach', args=(21, 33, 2)))
-                                     
+
+
 class PageDataTests(TestCase):
     def test_refresh_timer(self):
         'check refresh timer behavior'
@@ -535,4 +285,3 @@ class PageDataTests(TestCase):
         s = pageData.get_refresh_timer(request)
         self.assertNotEqual(s, '0:00')
         self.assertEqual(s[:3], '0:0')
-
