+++ mysite/ct/urls.py
@@ -132,9 +132,4 @@ urlpatterns = patterns('',
     url(r'^courses/$', courses, name='courses'),
     # Subscribe to course with particular id
     url(r'^courses/(?P<course_id>\d+)/subscribe/$', courses_subscribe, name='subscribe'),
-
-    # FSM node pages
-    url(r'^nodes/(?P<node_id>\d+)/$', fsm_node, name='fsm_node'),
-    url(r'^nodes/$', fsm_status, name='fsm_status'),
-
 )
