+++ mysite/fsm/fsm_plugin/fsmspec.py
@@ -1,7 +1,10 @@
-from ct.models import FSM
+from fsm.models import FSM
+
 
 class FSMSpecification(object):
-    'convenience class for specifying an FSM graph, loading it'
+    """
+    Convenience class for specifying an FSM graph, loading it.
+    """
     def __init__(self, name, title, nodeDict=None, edges=None,
                  pluginNodes=(), attrs=('help', 'path', 'data', 'doLogging'),
                  fsmGroups=(), **kwargs):
@@ -13,35 +16,52 @@ class FSMSpecification(object):
             nodeDict = {}
         if not edges:
             edges = []
-        for node in pluginNodes: # expect list of node class objects
+        for node in pluginNodes:  # expect list of node class objects
             modName = node.__module__.split('.')[-1]
             name = node.__name__
             d = dict(title=node.title, funcName=modName + '.' + name,
                      description=getattr(node, '__doc__', None))
-            for attr in attrs: # save node attributes
+            for attr in attrs:  # save node attributes
                 if hasattr(node, attr):
                     d[attr] = getattr(node, attr)
             nodeDict[name] = d
             for e in getattr(node, 'edges', ()):
-                e = e.copy() # prevent side effects
+                e = e.copy()  # prevent side effects
                 e['fromNode'] = name
                 edges.append(e)
         self.nodeData = nodeDict
         self.edgeData = edges
+
     def save_graph(self, *args, **kwargs):
-        'load this FSM specification into the database'
-        return FSM.save_graph(self.fsmData, self.nodeData, self.edgeData,
-                              fsmGroups=self.fsmGroups, *args, **kwargs)
-            
+        """
+        Load this FSM specification into the database.
+        """
+        return FSM.save_graph(
+            self.fsmData,
+            self.nodeData,
+            self.edgeData,
+            fsmGroups=self.fsmGroups,
+            *args,
+            **kwargs
+        )
+
+
 class CallerNode(object):
-    'base class for node representing a call to a sub-FSM'
+    """
+    Base class for node representing a call to a sub-FSM.
+    """
     def exceptCancel_edge(self, edge, fsmStack, request, **kwargs):
-        'implements default behavior: if sub-FSM cancelled, we cancel too'
-        fsmStack.pop(request, eventName='exceptCancel') # cancel this FSM
+        """
+        Implements default behavior: if sub-FSM cancelled, we cancel too.
+        """
+        fsmStack.pop(request, eventName='exceptCancel')  # cancel this FSM
         return edge.toNode
-    
-def deploy(modname, username, prefix='ct.fsm_plugin.'):
-    'load FSM specifications found in the specified plugin module'
+
+
+def deploy(modname, username, prefix='fsm.fsm_plugin.'):
+    """
+    Load FSM specifications found in the specified plugin module.
+    """
     import importlib
     mod = importlib.import_module(prefix + modname)
     l = []
@@ -49,9 +69,12 @@ def deploy(modname, username, prefix='ct.fsm_plugin.'):
         l.append(fsmSpec.save_graph(username))
     return l
 
+
 def deploy_all(username, ignore=('testme', '__init__', 'fsmspec'),
-               pattern='ct/fsm_plugin/*.py'):
-    'load all FSM specifications found via pattern but not ignore'
+               pattern='fsm/fsm_plugin/*.py'):
+    """
+    Load all FSM specifications found via pattern but not ignore.
+    """
     import glob
     import os.path
     l = []
