+++ mysite/fsm/fsm_plugin/live.py
@@ -1,11 +1,15 @@
-from ct.models import *
+from fsm.models import ActivityLog
+
 
 def quit_edge(self, edge, fsmStack, request, **kwargs):
-    'edge method that terminates this live-session'
+    """
+    Edge method that terminates this live-session.
+    """
     for studentState in fsmStack.state.linkChildren.all():
-        studentState.linkState = None # detach from our state
+        studentState.linkState = None  # detach from our state
         studentState.save()
     return edge.toNode
+
 
 QuitEdgeData = dict(
     name='quit', toNode='END', title='End this live-session',
@@ -15,31 +19,38 @@ QuitEdgeData = dict(
     showOption=True,
 )
 
+
 class START(object):
-    '''This activity will allow you to select questions
-    for students to answer in-class. '''
+    """
+    This activity will allow you to select questions
+    for students to answer in-class.
+    """
     def start_event(self, node, fsmStack, request, **kwargs):
         'event handler for START node'
         unit = fsmStack.state.get_data_attr('unit')
         course = fsmStack.state.get_data_attr('course')
         fsmStack.state.title = 'Teaching: %s' % unit.title
-        activity = ActivityLog(fsmName=fsmStack.state.fsmNode.fsm.name,
-                               course=course) # create a new activity
+        activity = ActivityLog(
+            fsmName=fsmStack.state.fsmNode.fsm.name,
+            course=course
+        )  # create a new activity
         activity.save()
         fsmStack.state.activity = activity
         fsmStack.state.isLiveSession = True
         return node.get_path(fsmStack.state, request, **kwargs)
     # node specification data goes here
-    path = 'ct:fsm_node'
+    path = 'fsm:fsm_node'
     title = 'Start Teaching a Live Session'
     edges = (
             dict(name='next', toNode='CHOOSE', title='Start asking a question',
                  showOption=True),
         )
 
+
 class CHOOSE(object):
-    '''At this step you choose a question to ask in this live
-    session'''
+    """
+    At this step you choose a question to ask in this live session.
+    """
     # node specification data goes here
     path = 'ct:unit_lessons'
     title = 'Choose a Question to Ask'
@@ -51,7 +62,8 @@ class CHOOSE(object):
                  help='''Click here to start posing this question to your
                  live session students.'''),
         )
-        
+
+
 class QUESTION(object):
     path = 'ct:live_question'
     title = 'Ask a question to students in a classroom live-session'
@@ -60,12 +72,13 @@ class QUESTION(object):
     Then click the START button and ask the students to think about
     the question for a minute or so, then briefly type whatever
     answer they come up with.  You will be able to monitor their
-    progress on this page in real-time.'''    
+    progress on this page in real-time.'''
     edges = (
         dict(name='next', toNode='ANSWER', title='Present the answer',
              help='''Click here to move to the assessment stage of this
              exercise. '''),
     )
+
 
 class ANSWER(object):
     quit_edge = quit_edge
@@ -80,21 +93,25 @@ class ANSWER(object):
         QuitEdgeData,
     )
 
+
 class RECYCLE(object):
-    '''You have completed presenting this question.  Do you want to
-    ask the students another question, or end this live session?'''
+    """
+    You have completed presenting this question.  Do you want to
+    ask the students another question, or end this live session?
+    """
     def next_edge(self, edge, fsmStack, request, pageData=None, **kwargs):
         'make sure timer is reset before going to another question'
         pageData.set_refresh_timer(request, False)
         return edge.toNode
-    path = 'ct:fsm_node'
+    path = 'fsm:fsm_node'
     title = 'Do you want to ask another question?'
     edges = (
         dict(name='next', toNode='CHOOSE', title='Move on to another question',
              help='''Click here to choose another question to ask. '''),
         QuitEdgeData,
     )
-        
+
+
 class END(object):
     # node specification data goes here
     path = 'ct:unit_tasks'
@@ -103,14 +120,16 @@ class END(object):
     See below for suggested next steps for what you can work on next
     to help students with this courselet.'''
 
+
 def get_specs():
     'get FSM specifications stored in this file'
     from fsmspec import FSMSpecification
-    spec = FSMSpecification(name='liveteach', 
-            title='Teach a live (classroom) session',
-            description='''You can begin teaching this courselet in a
-            live classroom session by clicking here:''',
-            pluginNodes=[START, CHOOSE, QUESTION, ANSWER, RECYCLE, END],
-            fsmGroups=('teach/unit/published',),
-        )
+    spec = FSMSpecification(
+        name='liveteach',
+        title='Teach a live (classroom) session',
+        description='''You can begin teaching this courselet in a
+                    live classroom session by clicking here:''',
+        pluginNodes=[START, CHOOSE, QUESTION, ANSWER, RECYCLE, END],
+        fsmGroups=('teach/unit/published',),
+    )
     return (spec,)
