+++ mysite/fsm/fsm_plugin/livestudent.py
@@ -1,36 +1,41 @@
-from ct.models import *
-
 def ask_edge(self, edge, fsmStack, request, **kwargs):
-    'try to transition to ASK, or WAIT_ASK if not ready'
+    """
+    Try to transition to ASK, or WAIT_ASK if not ready.
+    """
     fsm = edge.fromNode.fsm
-    if not fsmStack.state.linkState: # instructor detached
+    if not fsmStack.state.linkState:  # instructor detached
         return fsm.get_node('END')
-    elif fsmStack.state.linkState.fsmNode.name == 'QUESTION': # in progress
+    elif fsmStack.state.linkState.fsmNode.name == 'QUESTION':  # in progress
         fsmStack.state.unitLesson = fsmStack.state.linkState.unitLesson
         fsmStack.state.save()
-        return edge.toNode # so go straight to asking question
+        return edge.toNode  # so go straight to asking question
     return fsm.get_node('WAIT_ASK')
 
 
 def assess_edge(self, edge, fsmStack, request, **kwargs):
-    '''try to transition to ASSESS, or WAIT_ASSESS if not ready,
-    or jump to ASK if a new question is being asked.'''
+    """
+    Try to transition to ASSESS, or WAIT_ASSESS if not ready,
+    or jump to ASK if a new question is being asked.
+    """
     fsm = edge.fromNode.fsm
-    if not fsmStack.state.linkState: # instructor detached
+    if not fsmStack.state.linkState:  # instructor detached
         return fsm.get_node('END')
     elif fsmStack.state.linkState.fsmNode.name == 'QUESTION':
         if fsmStack.state.unitLesson == fsmStack.state.linkState.unitLesson:
             return fsm.get_node('WAIT_ASSESS')
-        else: # jump to the new question
+        else:  # jump to the new question
             fsmStack.state.unitLesson = fsmStack.state.linkState.unitLesson
             fsmStack.state.save()
             return fsm.get_node('ASK')
     else:
-        return edge.toNode # go to assessment
+        return edge.toNode  # go to assessment
+
 
 class START(object):
-    '''In this activity you will answer questions
-    presented by your instructor in-class. '''
+    """
+    In this activity you will answer questions
+    presented by your instructor in-class.
+    """
     def start_event(self, node, fsmStack, request, **kwargs):
         'event handler for START node'
         fsmStack.state.activity = fsmStack.state.linkState.activity
@@ -42,26 +47,32 @@ class START(object):
         return node.get_path(fsmStack.state, request, **kwargs)
     next_edge = ask_edge
     # node specification data goes here
-    path = 'ct:fsm_node'
+    path = 'fsm:fsm_node'
     title = 'Now Joining a Live Classroom Session'
     edges = (
         dict(name='next', toNode='ASK', title='Start answering questions'),
     )
 
+
 class WAIT_ASK(object):
-    '''The instructor has not assigned the next exercise yet.
+    """
+    The instructor has not assigned the next exercise yet.
     Please wait, and click the Next button when the instructor tells
-    you to do so, or when the live classroom session is over. '''
+    you to do so, or when the live classroom session is over.
+    """
     next_edge = ask_edge
     # node specification data goes here
-    path = 'ct:fsm_node'
+    path = 'fsm:fsm_node'
     title = 'Wait for the Instructor to Assign a Question'
     edges = (
         dict(name='next', toNode='ASK', title='See if question assigned'),
     )
 
+
 class ASK(object):
-    '''In this stage you write a brief answer to a conceptual question. '''
+    """
+    In this stage you write a brief answer to a conceptual question.
+    """
     def next_edge(self, edge, fsmStack, request, response=None, **kwargs):
         if response:
             fsmStack.state.set_data_attr('response', response)
@@ -80,20 +91,26 @@ class ASK(object):
         dict(name='next', toNode='ASSESS', title='Proceed to assessment'),
     )
 
+
 class WAIT_ASSESS(object):
-    '''The instructor has not ended the question period yet.
+    """
+    The instructor has not ended the question period yet.
     Please wait, and click the Next button when the instructor tells
-    you to do so, or when the live classroom session is over. '''
+    you to do so, or when the live classroom session is over.
+    """
     next_edge = assess_edge
     # node specification data goes here
-    path = 'ct:fsm_node'
+    path = 'fsm:fsm_node'
     title = 'Wait for the Instructor to End the Question'
     edges = (
         dict(name='next', toNode='ASSESS', title='See if question done'),
     )
 
+
 class ASSESS(object):
-    '''In this stage you assess your own answer vs. the correct answer '''
+    """
+    In this stage you assess your own answer vs. the correct answer.
+    """
     next_edge = ask_edge
     error_edge = ask_edge
     # node specification data goes here
@@ -107,7 +124,7 @@ class ASSESS(object):
         dict(name='error', toNode='ASK', title='Go to the next question'),
     )
 
-        
+
 class END(object):
     # node specification data goes here
     path = 'ct:unit_tasks_student'
@@ -116,12 +133,16 @@ class END(object):
     See below for suggested next steps for what to study now in
     this courselet.'''
 
-        
+
 def get_specs():
-    'get FSM specifications stored in this file'
+    """
+    Get FSM specifications stored in this file.
+    """
     from fsmspec import FSMSpecification
-    spec = FSMSpecification(name='livestudent', hideTabs=True,
-            title='Join a Live Classroom Session',
-            pluginNodes=[START, WAIT_ASK, ASK, WAIT_ASSESS, ASSESS, END],
-        )
+    spec = FSMSpecification(
+        name='livestudent',
+        hideTabs=True,
+        title='Join a Live Classroom Session',
+        pluginNodes=[START, WAIT_ASK, ASK, WAIT_ASSESS, ASSESS, END],
+    )
     return (spec,)
