+++ mysite/templates/fsm/fsm_node.html
