+++ mysite/templates/fsm/fsm_status.html
