+++ prod_requirements.txt
@@ -33,3 +33,4 @@ requests==2.7.0
 requests-oauthlib==0.5.0
 six==1.9.0
 wikipedia==1.4.0
+sqlparse==0.1.15
