+++ requirements.txt
@@ -48,3 +48,4 @@ pep8==1.6.2
 pylint==1.4.3
 ddt==1.0.0
 Sphinx==1.3.1
+sqlparse==0.1.15
